requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.1
Django==4.1.6
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
pytz==2022.7.1
sqlparse==0.4.3
